backend/requirements.txt
# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
prisma==0.15.0
asyncpg==0.29.0

# HTTP and multipart support
python-multipart==0.0.6
httpx==0.25.2

# Configuration
python-decouple==3.8
python-dotenv==1.1.1

# Image processing
pillow==10.1.0
opencv-python==4.8.1.78
pytesseract==0.3.10

# Data processing
pandas==2.1.3
numpy==1.26.4

# Data validation
pydantic==2.5.0
pydantic-core==2.14.1

# AI services
anthropic==0.61.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Additional dependencies (auto-installed)
anyio==3.7.1
async-timeout==5.0.1
starlette==0.27.0
greenlet==3.2.3
typing-extensions==4.14.1